Split Home and Article examples
