--- storybook/src/pages/amsterdam/amsterdam.docs.mdx
@@ -1,26 +1,0 @@
-import { Meta } from "@storybook/blocks";
-import * as AmsterdamStories from "./amsterdam.stories.tsx";
-
-<Meta of={AmsterdamStories} />
-
-# Amsterdam.nl Pattern Library
-
-## Hompage
-
-The homepage of Amsterdam.nl is the first page that users see when they visit the website. It is the most important page of the website, as it is the first impression that users get of the website. The homepage is designed to be visually appealing and easy to navigate, with a clear layout and intuitive navigation.
-
-```tsx
-<SkipLink href="#main">Direct naar inhoud</SkipLink>
-<Screen maxWidth="wide">
-  <Grid>
-    <Header />
-  </Grid>
-  <Overlap>
-    <Search />
-  </Overlap>
-  <Grid>
-    <MostVisited />
-  </Grid>
-  <Footer />
-</Screen>
-```
