+++ storybook/src/pages/amsterdam/article/Article.docs.mdx
@@ -1,0 +1,8 @@
+import { Meta } from "@storybook/blocks";
+import * as ArticleStories from "./Article.stories.tsx";
+
+<Meta of={ArticleStories} />
+
+# Article Page
+
+TODO
